Fix Wildcard Resource Rule test that was failing from a malformed template
This is not passing pydantic validation due to not having a PolicyName
property. However, this was by design, so I have kept the test and have it
test for the error.

There exists an equivalent test that has the PolicyName property, so no test
seemingly missing here by not having it.

See here https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-iam-policy.html
PolicyName is `Required: Yes` as it is in the pycfmodel too
https://github.com/Skyscanner/pycfmodel/blob/master/pycfmodel/model/resources/properties/policy.py#L16